move HydraContext to hydra.types
