+++ examples/plugins/example_sweeper_plugin/hydra_plugins/example_sweeper_plugin/example_sweeper.py
@@ -6,7 +6,7 @@ import logging
 from pathlib import Path
 from typing import Any, Iterable, List, Optional, Sequence
 
-from hydra.core.utils import HydraContext
+from hydra.types import HydraContext
 from hydra.core.config_store import ConfigStore
 from hydra.core.override_parser.overrides_parser import OverridesParser
 from hydra.core.plugins import Plugins
