+++ plugins/hydra_joblib_launcher/hydra_plugins/hydra_joblib_launcher/joblib_launcher.py
@@ -2,7 +2,8 @@
 import logging
 from typing import Any, Optional, Sequence
 
-from hydra.core.utils import HydraContext, JobReturn
+from hydra.core.utils import JobReturn
+from hydra.types import HydraContext
 from hydra.plugins.launcher import Launcher
 from hydra.types import TaskFunction
 from omegaconf import DictConfig
