+++ plugins/hydra_ray_launcher/hydra_plugins/hydra_ray_launcher/ray_launcher.py
@@ -1,7 +1,8 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 from typing import Optional, Sequence
 
-from hydra.core.utils import HydraContext, JobReturn
+from hydra.core.utils import JobReturn
+from hydra.types import HydraContext
 from hydra.plugins.launcher import Launcher
 from hydra.types import TaskFunction
 from omegaconf import DictConfig
